agentic-rag-aws/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
chromadb
sentence-transformers
pypdf